requirements.txt
Django~=4.2.1
backports.zoneinfo==0.2.1;python_version<"3.9"
djangorestframework~=3.14.0
django-colorfield ~= 0.8.0
psycopg2-binary~=2.9.6
django-faker~=0.2
django-filter~=23.2
